Ignore .qws file
A Quartus .qws file is a non-editable file that stores user preferences and other information related to the Quartus Prime software workspace. This includes the position of windows, open files, window docking status, and other settings. The file is automatically generated by Quartus Prime and is used to maintain the user's working environment.